+++ CHANGELOG.md
@@ -8,6 +8,8 @@
     - `Action.mute/check/reject/then` will take `AcceptedActionOutput` as the callback's param.
     - Regex should not starts with `^` and will auto have sticky flag.
   - **_Breaking Change_**: `Lexer.whitespaces` is now a function.
+  - **_Breaking Change_**: add `ILexer.lineChars/errors`, remove `ILexer.getLineChars/getErrors`.
+  - Feat: add `ILexer.defs/buffer`.
   - Feat: `SimpleActionExec` can accept `string/SimpleAcceptedActionOutput` as the output.
   - Feat: add generic parameter for Lexer family as the `Token.error` type.
   - Feat: add `options.autoSticky/autoGlobal` for `fromTo`.